refactor: decompose DataInitializer into DataSyncContributor beans (ATW-8f45.2)
Reduces DataInitializer.java from 1,871 lines (health score 1.0/10) to ~50 lines by
extracting each syncXxxFromFile() method into its own @Component implementing
DataSyncContributor. Spring auto-orders contributors via @Order(10..250) and wires them
as List<DataSyncContributor>. Adds characterization tests for AiSettingsSync, NpcSync,
and WrestlerSync; rewrites DataInitializerTest to exercise the thin orchestrator.
